refactor: Update resource authorization mode and default authorization mode
This commit updates the resource authorization mode in the `amplify/data/resource.ts` file. The previous authorization mode was set to `allow.publicApiKey()`, but it has been changed to `allow.owner()`.

Additionally, the default authorization mode in the same file has been updated from `"apiKey"` to `"userPool"`.

These changes were made to improve the security and access control of the resource.